change from u32 to u64

+++ src/components/EatAnApple.tsx
@@ -36,7 +36,7 @@ async function eatAnApple () {
         {
           gasLimit: gasLimit,
           storageDepositLimit,
-        }, {u32: token_number}, account.address)
+        }, {u64: token_number}, account.address)
 
         console.log("### result of dry run ###" );
         console.log("### output:", output?.toHuman());
@@ -56,7 +56,7 @@ async function eatAnApple () {
             {
               gasLimit: gasLimit,
               storageDepositLimit,
-            }, {u32: token_number}, account.address).signAndSend(account.address, { signer: injector.signer }, async ({ status }) => {
+            }, {u64: token_number}, account.address).signAndSend(account.address, { signer: injector.signer }, async ({ status }) => {
     
               if (status.isInBlock) {
                   console.log(`Completed at block hash #${status.asInBlock.toString()}`);
